+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 Th 17-03-2016
+Haskier a-0.2
+- New commands : set, force-save
+- Added game loader (com/get-game.run)
+- Game can now be saved
+- Created server feature, including import/export
+- Support of hidden files
+- HSF (Haskier Scenario File) are now supported [Math.js 1.7.1 parser]
+*** === Bug fixes === ***
+- Server was returning bad file length for plain text
+
+Th 17-03-2016
 Haskier a-0.1
 - Commands are supported
 - Support of asynchronous commands